feat: basicInput, basicSearch 공통 interface 정의
