+++ src/types/components/BasicInputProps.ts
@@ -1,0 +1,11 @@
+import { MutableRefObject } from 'react';
+
+export interface BasicInputProps {
+  type?: string;
+  label: string;
+  isRequired?: boolean;
+  defaultValue?: string;
+  regExp?: string;
+  errorMessage?: string;
+  inputRef: MutableRefObject<object>;
+}
